WIP Creation feature

+++ app/src/main/java/com/example/loan_online/features/create/data/ConditionsDTO.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ConditionsDTO(
 
-    @SerializedName("maxAmount") val maxAmount: Int,
-    @SerializedName("percent") val percent: Int,
+    @SerializedName("maxAmount") val maxAmount: Double,
+    @SerializedName("percent") val percent: Double,
     @SerializedName("period") val period: Int
 )
