+++ app/src/main/java/com/example/loan_online/features/create/domain/ConditionsModel.kt
@@ -2,7 +2,7 @@ package com.example.loan_online.features.create.domain
 
 data class ConditionsModel(
 
-    val maxAmount: Int,
-    val percent: Int,
+    val maxAmount: Double,
+    val percent: Double,
     val period: Int
 )
